Give deep dwarves halved healing.
This is an attempt at balancing deep dwarves that should address
both the overpowerment of /healing + innate recharging and
deep dwarves with other cheap ways of healing.

The factor 1/2 is obviously subject to tweaking: It could be
changed, or scaled over the game. For example, It could start
at 1 and end at 1/3 at XL 27 if early game is deemed hard
enough currently.

If we want to go with this, it should probably become a racial
mutation, but for testing the current minimal change should be
sufficient.