Explosive Thaw: L5 Fire/Ice
Range 2 fireball with lower, typeless damage, which also covers
the affected squares with shallow water for a few (2-7) turns.
Good for messing with melee monsters, possibly?

Might be better as Fire/Ice/Conj (with full fireball damage?) -
I have a vague feeling that there are too many Cj spells already,
but the triple-school approach makes it more of a "can be used by
IE or FE" spell, rather than "train both ice and fire to use this".
Which... well, I'm not sure what's better.

The theming is a mess. Sad!