New spell: Reckless Fragmentation (L7 Conj/Earth)
An attempt to experiment with different targetting types & to
expand on Earth's use of fragmentation (3x AC-resisted) damage.

The spell is smite-targeted (range 4) and causes a radius 2
explosion somewhere within 2 tiles of the target. That means that
it will always hit the targetted square, but adjacent tiles are
less likely to be hit, and those one further away less likely yet.
High AC is recommended (statue form/'armour caster' synergy...)

Current damage is about 3/4ths of Firestorm's, which isn't nearly
as strong as it sounds once 3x AC kicks in. I'm not committed to
that number, though. (Nor to the current radius - a radius 1
explosion offset by 1 from the target could work, though it seems
less interesting to me.)

Right now it's just in the Book of Earth - could be a decent
Annihilations option.

XXX: Suppress "really fire at yourself", at least without veh's
range extension?