Re-organize species selection screen
The old approach was to use fantasy taxonomy ('comparatively
human-like looks', 'monstrous', 'humanoid but made of a strange
substance', etc), which was neither very obvious nor very useful to
the player. It seems more useful to try to (very roughly!) sort
races by 'appropriateness for new players', from strongest & simplest
to hardest & most complex.

To make this clearer, it could be worthwhile to add section headers,
as with the background selection screen.