Add a way of minimizing and hiding the chat window
This adds a third mode to the chat window: it can be minimized into a
small box with just a `+` that does not visually change with chat state,
in case (for any number of reasons) a player doesn't want to be
distracted by chat. The window can be restored by clicking the +, or by
hitting F12.

There's also a slightly hidden mode that really gets rid of chat, aimed
at streamers who are currently just doing this with browser dev tools.
To get this, type `/hide forever` in the chat window. To get it back,
you will need to reload, or log out and in.

None of this state is saved, although it could be if there's demand for
that.