Nerf deep troll armour and troll iron armour
This commit makes deep troll leather armour give 6 AC for 7 ER, to be
better than ring mail (5/7) but worse than acid dragon scales (6/5) and
swamp dragon scales (7/7); and iron troll armour give 8 AC for 13 ER, to
be better than chain mail but worse than the various dragon scales. This
is to make them not-strictly-better than the dragon scales when
resistances are not required.