Make the Velocity card less harmful at low power
At power level zero Velocity was far more likely to be actively harmful
to the player than to help. With the changes to Nemelex, this particular
blind-draw downside is too harsh.

The new card no longer removes haste at power zero, and has the
following effects based on card level.

- level chance in 2 to remove slow from the player
- for each enemy in sight there is a (level + 1) chance in 3 to slow
- for each ally in sight there is a level chance in 2 to haste