Display the top of the stack in the ability menu
Currently 'the Pentagram' is 13 characters, so there isn't space in the
Cost column of the ability menu for the string
'A Card (the Pentagram, n more)' using a shorter string than this makes
it even harder to parse. Since a player concerned about how many cards
remain in their stack is probably also interested in which cards those
are it doesn't seem too bad to not have the count in the ability
display.