Use contribs if appropriate when building rltiles/ via sub-make
This passes along the NO_PKGCONFIG value determined by the main Makefile
to rltiles/Makefile. Previously, the sub-make process would use
pkg-config libpng regardless of what the main Makefile is doing. This
isn't necessarily a big deal, because the utility built with this isn't
distributed, but it is breaking the build for hypothetical someone who
is having pkg-config problems and wants to rely just on contribs (me),
or someone who might try to build crawl in a restricted environment
without pkg-config (which apparently is uncommon).