Fix surge of power message order when evoking enhanced wands (11753)
This commit adds functionality to deal with the "surge of power" message
when zapping wands under MP-powered wands in your_spells directly, so
that this message is displayed before any of the wand's effects and any
world_reacts effects caused by zapping the wand are displayed.

To effect the change this commit also moves the check for the player
having enough MP to use the full power bonus from MP-powered wands into
wand_mp_cost.

[Committer's note: Closes #926, squashed and kept some Pakellas code in
place since Pak is disabled but not removed]