Fix out-of-order messages (maybe 11780)
For compressed websockets, the decoding process used a FileReader to
convert the output of decompression to UTF-8. However, FileReader comes
with no guarantees about the order in which callbacks get attended to
when there are multiple instances, and in chrome especially they could
get quite out of order for messages sent at roughly the same time. This
lead to the message queue being out-of-order, which various aspects of
the webtiles client are not robust to.

This commit uses (suggested by jilles) a TextDecoder instance instead.
Because this may not be supported by some browser versions still in use,
I have kept the old approach as a fallback. (An alternative would be
simply to use uncompressed websockets if TextDecoder is not supported.)