Stop dragon cruelty in a vault (howmanybefore)
Dragons were sealed behind clear stone walls in orc_legates. Items could
generate there and it was a teleport closet. We could mask this area
with no_tele_into and no_item_gen, but it's more interesting to add
clear runed doors and spruce up the dragon collection a bit, so do that
instead.