Further distinguish light skill drain
This should make it easier for the player to judge how heavily he
is drained in the early game, and when using Ru's powers.

For monsters, the drain of a wraith is 25, and a hit from a draining
wand is a minimum of 75, but currently both give you the same status.

The same is true for Ru's Draw out power and Apocalypse: 30 vs 100,
but both give you the same status (and have the same "skill drain"
description in the god ability menu). For Ru it is quite important to
manage your skill drain, and introducing a new threshold should help
with that.

[Committer's note: Adjusted the colour scale; Closes #963]