Include petrification information in death messages (11374)
Display "Slain by X ... while petrified [by X]" on the death screen when
the player was killed while petrified, in the same way as paralysis.

This applies only when fully petrified, not when petrifying, and the
monster that first applied the petrifying status is used in the death
message, even if the player is hit again with the petrify spell by a
different source on the following turn and this accelerated the
petrification process.