Remove crawl.random_element and replace with something more stable
Because lua iteration order, including at the C level, is not stable,
this was leading to seed variation in parts of the dungeon that heavily
use layout code. I've replaced it with a pure lua implementation of
random_choose_weighted (that parallels the C++ version); in principle
one could call into the C++ version, but it wasn't clear that there was
much of a point to that amount of effort. This version is stable because
it uses an array of pairs for the weight table; `ipairs` gives a stable
iteration order. Notationally it's not any more effortful.