Reset temple vault tags when doing temple setup
This load() call was just adding in new tags to the previously loaded
list each time a game was started, leading to seed instability both in
tests and when a new game was started without exiting the program for
temple vaults with variable numbers of altars. (Because the length of
the list of tags changed, which leads to different behavior of the
random draw as the list gets longer even for the same seed.)

Possibly tags should be treated as a `set` or something, rather than
just a raw string that is split on demand?