Invalidate LOS properly for level pregeneration
It turns out that monster band placement checks LOS when placing the
followers, and so if LOS is not properly invalidated, the builder will
use the cache for the wrong level, leading to seed divergence. This
moves a few other cleanups up that I think will matter a bit less.