Fix a couple of des cache random tag bugs
Because lua code is run during des cache validation, any random
decisions made during validation may be incorporated into the cache. I
*think* this is mostly harmless but there are at least two problems that
arise when it applies to vaults tags. First, random choices when
selecting maps can be impacted by the number of tags *in the cache*, so
seeds will potentially behave differently depending on choices made when
first building the cache. Second, because tags are strictly additive, a
tag choice made during cache building will never get removed; for
example in `lemuel_altar_in_water` before this commit, if the cache
rolled no_pool_fixup, this would be present in all instances of this
vault regardless of rolls when actually placing the vault.

It's possible that tags should be cleared when rerunning vault lua for
placement, but that seems like a change that could potentially impact a
lot of things and will require some testing, so I'm going to postpone
it.

In general it would probably be best to eliminate *all* randomization
from the cache, but I'm not yet sure what the best technical solution to
that is (there are quite a few vaults that use random calls like this to
set SUBST or monster values, it doesn't seem ideal to impose
is_validating checks on all of them).