Fix Beogh crash for polymorphed followers (11880 / reaverb)
If an orc gets turned into something that cannot hold items (e.g. via a
chaos cloud shapeshifter effect) it will drop its gift, but the gift
prop does not get cleared, something the new ^ extra screen was not
checking for. This commit fixes things at the point of the crash; it's
possible that the prop should get cleared when the item is dropped in
the first place.