Don't try to describe off-level orc gifts (11881)
Instead just say that the orc is off-level. Better would be to describe
both the item and where they are, but the the way off-level followers
are implemented means that it's basically impossible to get the item
information without either a level excursion or a major rewrite.

One option might be to save the item string as a prop?