Clean up god gifts' interaction with delays (11825)
Getting a Nemelex gift at the end of a successful passwall that revealed
new cells caused the message "Your meditation is interrupted." to be
printed. This is because do_god_gift interrupted running and resting in
a way that interrupted and removed the passwall delay. We now call
stop_running with the clear_delays parameter set to false in order to no
longer clear delays. We also call stop_running instead of directly
reaching down to you.running.stop.