Improve behavior of local tiles Text, especially on hidpi displays
This commit ends up doing a number of things:

1. (most important) fix a bug on retina (and other hidpi) displays where
the height of Text widget contents was not calculated correctly, because
the code didn't handle fractional y advances. This resulted in ctrl-p
dropping several lines at the end, among other things.
2. Fix handling of fractional y advances when drawing highlights, again
a potential issue in long Texts (though less important, as highlighting
is almost never used right now).

In the course of this, I needed to understand how the highlight code
worked and this forced me to add a lot of comments and do some code
formatting, as well as address the following bugs:

3. Don't highlight lines more than once in hl_line mode, this results
in unreadable text.
4. Handle the case where the end of a string coincides exactly with the
end of the window and is highlighted, which was leading to glitches in
my testing.
5. Better handling for cases where a highlight spans multiple slices of
the formatted string. I'm not 100% sure if this was really broken before
my fix of 4, but it works now.

I'm still not entirely sure about all this highlighting code, and it's
pretty complicated and monolithic, so it could probably use some more
work and refactoring.